@@ -3,45 +3,32 @@ import com.techreturners.bubbleteaordersystem.service.BubbleTeaRouletteService;
 
 import testhelper.RandomStub;
 
-import org.junit.Test;
-import org.junit.runner.RunWith;
-import org.junit.runners.Parameterized;
-import org.junit.runners.Parameterized.Parameters;
+import java.util.stream.Stream;
 
-import java.util.Arrays;
-import java.util.Collection;
+import org.junit.jupiter.params.ParameterizedTest;
+import org.junit.jupiter.params.provider.Arguments;
+import org.junit.jupiter.params.provider.MethodSource;
+import org.junit.jupiter.api.Assertions;
 
-import static org.junit.Assert.assertEquals;
-
-@RunWith(Parameterized.class)
 public class BubbleTeaRouletteServiceParameterizedTest {
 
-    @Parameters
-    public static Collection<Object[]> data() {
-        return Arrays.asList(new Object[][]{
-                {BubbleTeaTypeEnum.PeachIceTea.ordinal(), BubbleTeaTypeEnum.PeachIceTea},
-                {BubbleTeaTypeEnum.JasmineMilkTea.ordinal(), BubbleTeaTypeEnum.JasmineMilkTea},
-                {BubbleTeaTypeEnum.MatchaMilkTea.ordinal(), BubbleTeaTypeEnum.MatchaMilkTea},
-                {BubbleTeaTypeEnum.PeachIceTea.ordinal(), BubbleTeaTypeEnum.PeachIceTea},
-                {BubbleTeaTypeEnum.LycheeIceTea.ordinal(), BubbleTeaTypeEnum.LycheeIceTea}
-        });
-    }
-
-    private final int input;
-    private final BubbleTeaTypeEnum resultExpected;
-
-    public BubbleTeaRouletteServiceParameterizedTest(final int input, final BubbleTeaTypeEnum result) {
-        this.input = input;
-        this.resultExpected = result;
-    }
-
-    @Test
-    public void testBubbleTeaRouletteService() {
+    @ParameterizedTest
+    @MethodSource("bubbleTeaTypes")
+    void testBubbleTeaRouletteService (int input, BubbleTeaTypeEnum resultExpected)
+    {
         RandomStub randomStub = new RandomStub(input);
         BubbleTeaRouletteService bubbleTeaRouletteService = new BubbleTeaRouletteService(randomStub);
-        assertEquals(resultExpected, bubbleTeaRouletteService.getRandomBubbleTeaType());
+        Assertions.assertEquals(resultExpected, bubbleTeaRouletteService.getRandomBubbleTeaType());
     }
 
+    private static Stream<Arguments> bubbleTeaTypes() {
+        return Stream.of(
+                Arguments.of(BubbleTeaTypeEnum.OolongMilkTea.ordinal(), BubbleTeaTypeEnum.OolongMilkTea),
+                Arguments.of(BubbleTeaTypeEnum.JasmineMilkTea.ordinal(), BubbleTeaTypeEnum.JasmineMilkTea),
+                Arguments.of(BubbleTeaTypeEnum.MatchaMilkTea.ordinal(), BubbleTeaTypeEnum.MatchaMilkTea),
+                Arguments.of(BubbleTeaTypeEnum.PeachIceTea.ordinal(), BubbleTeaTypeEnum.PeachIceTea),
+                Arguments.of(BubbleTeaTypeEnum.LycheeIceTea.ordinal(), BubbleTeaTypeEnum.LycheeIceTea));
+    }
 
 }
 
